Add Status Bar configuration via Cordova Plugin
Configure the Status Bar with plugin options hardcoded. Phonegap will
do the right thing on its own with the values for mobile devices, and
does not involve Javascript or CSS "trickery".

The issue with this change is the background colour is set to black
with no way to make it transparent, or match the background of the
application.

Note, this change may require a new IPA as Phonegap might need it to
handle the config.xml changes.